+++ HW07/task2.cu
@@ -9,28 +9,28 @@ int main(int argc, char* argv[]) {
     int t = atoi(argv[2]);
     int blocks = (n + t - 1) / t / 2;
     
-    int hInput[n], hOutput[blocks], *dInput, *dOutput;
+    float hInput[n], hOutput[blocks], *dInput, *dOutput;
 
     // randomization:
     srand(chrono::system_clock::now().time_since_epoch().count());
 
     // array initialization:
     for (int i = 0; i < n; i++){
-        hInput[i] = static_cast <int> (rand() / static_cast <int> (RAND_MAX / 2)) - 1;
+        hInput[i] = static_cast <float> (rand() / static_cast <float> (RAND_MAX / 2)) - 1;
     }
 
     auto start = chrono::steady_clock::now();
 
-    cudaMalloc((void**)&dInput, sizeof(int) * n);
-    cudaMalloc((void**)&dOutput, sizeof(int) * blocks);
-    cudaMemcpy(dInput, &hInput, sizeof(int) * n, cudaMemcpyHostToDevice);
-    cudaMemset(dOutput, 0, sizeof(int) * blocks);
+    cudaMalloc((void**)&dInput, sizeof(float) * n);
+    cudaMalloc((void**)&dOutput, sizeof(float) * blocks);
+    cudaMemcpy(dInput, &hInput, sizeof(float) * n, cudaMemcpyHostToDevice);
+    cudaMemset(dOutput, 0, sizeof(float) * blocks);
 
     // do math:
     reduce(&dInput, &dOutput, n, t);
 
     // results:
-    cudaMemcpy(&hOutput, dOutput, sizeof(int) * blocks, cudaMemcpyDeviceToHost);
+    cudaMemcpy(&hOutput, dOutput, sizeof(float) * blocks, cudaMemcpyDeviceToHost);
 
     auto end = chrono::steady_clock::now();
 	auto timePassed = chrono::duration_cast<std::chrono::microseconds>(end - start);
